restructure sidechain_storage file into folder
